requirements.txt
setuptools
webdataset
pytorch-lightning
tensorboard
typer
psutil
wandb
webdataset
ray[default]
ray[air]
ipython
jupyterlab
